+++ src/flows.py
@@ -17,13 +17,10 @@ from config import (
     MAX_PREFECT_WORKERS,
     SIRENE_DATA_DIR,
     TRACKED_DATASETS,
-    make_dirs_if_not_exist,
 )
-from tasks.analyse import generate_stats
-from tasks.cache_management import remove_unused_cache
 from tasks.clean import clean_decp
 from tasks.dataset_utils import list_resources
-from tasks.enrich import add_unite_legale_data
+from tasks.enrich import enrich_from_sirene
 from tasks.get import get_resource
 from tasks.output import (
     save_to_files,
@@ -32,13 +29,11 @@ from tasks.output import (
 from tasks.publish import publish_to_datagouv
 from tasks.transform import (
     concat_decp_json,
-    extract_unique_acheteurs_siret,
-    extract_unique_titulaires_siret,
     get_prepare_unites_legales,
-    make_decp_sans_titulaires,
     normalize_tables,
     sort_columns,
 )
+from tasks.utils import create_sirene_data_dir, generate_stats, remove_unused_cache
 
 
 @task(
@@ -62,16 +57,16 @@ def get_clean(resource):
     return lf
 
 
-@flow(log_prints=True)
+@task(log_prints=True)
 def make_datalab_data():
     """Tâches consacrées à la transformation des données dans un format
     adapté aux activités du Datalab d'Anticor."""
 
-    print("🚀  Création des données pour le Datalab d'Anticor...")
+    print("Création de la base données pour le Datalab d'Anticor...")
 
     df: pl.DataFrame = pl.read_parquet(DIST_DIR / "decp.parquet")
 
-    print("Enregistrement des DECP aux formats SQLite...")
+    print("Enregistrement des DECP (base DataFrame) au format SQLite...")
     save_to_sqlite(
         df,
         "datalab",
@@ -88,50 +83,12 @@ def make_datalab_data():
     else:
         print("Publication sur data.gouv.fr désactivée.")
 
-    print("☑️  Fin du flow make_datalab_data.")
-
-
-@flow(log_prints=True)
-def make_decpinfo_data():
-    """Tâches consacrées à la transformation des données dans un format
-    # adapté à decp.info"""
-
-    print("🚀  Création des données pour decp.info...")
-
-    df: pl.DataFrame = pl.read_parquet(DIST_DIR / "decp.parquet")
-
-    # DECP sans titulaires
-    save_to_files(make_decp_sans_titulaires(df), DIST_DIR / "decp-sans-titulaires")
-
-    # print("Ajout des colonnes manquantes...")
-    # df = setup_tableschema_columns(df)
-
-    # CREATION D'UN DATA PACKAGE (FRICTIONLESS DATA)
-
-    # Pas la priorité pour le moment, prend du temps
-    # print("Validation des données DECP avec le TableSchema...")
-    # validate_decp_against_tableschema()
-
-    # print("Création du data package (JSON)....")
-    # make_data_package()
-
-    # PUBLICATION DES FICHIERS SUR DATA.GOUV.FR
-    if DECP_PROCESSING_PUBLISH.lower() == "true":
-        print("Publication sur data.gouv.fr...")
-        publish_to_datagouv(context="decp")
-    else:
-        print("Publication sur data.gouv.fr désactivée.")
-
-    print("☑️  Fin du flow make_decpinfo_data.")
-
-    return df
-
 
 @flow(
     log_prints=True, task_runner=ConcurrentTaskRunner(max_workers=MAX_PREFECT_WORKERS)
 )
 def decp_processing(enable_cache_removal: bool = False):
-    print("🚀  Début du flow principal")
+    print("🚀  Début du flow decp-processing")
 
     print("Liste de toutes les ressources des datasets...")
     resources: list[dict] = list_resources(TRACKED_DATASETS)
@@ -144,6 +101,11 @@ def decp_processing(enable_cache_removal: bool = False):
     df: pl.DataFrame = concat_decp_json(lfs)
 
     print("Ajout des données SIRENE...")
+    # Preprocessing des données SIRENE si :
+    # - le dossier n'existe pas encore (= les données n'ont pas déjà été preprocessed ce mois-ci)
+    # - on est au moins le 5 du mois (pour être sûr que les données SIRENE ont été mises à jour sur data.gouv.fr)
+    if not SIRENE_DATA_DIR.exists() and int(DATE_NOW[-2:]) >= 5:
+        sirene_preprocess()
     lf: pl.LazyFrame = enrich_from_sirene(df.lazy())
 
     print("Génération de l'artefact (statistiques) sur le base df...")
@@ -159,9 +121,6 @@ def decp_processing(enable_cache_removal: bool = False):
     df: pl.DataFrame = sort_columns(df, BASE_DF_COLUMNS)
     save_to_files(df, DIST_DIR / "decp")
 
-    # Fichiers dédiés à l'Open Data et decp.info
-    make_decpinfo_data()
-
     # Base de données SQLite dédiée aux activités du Datalab d'Anticor
     make_datalab_data()
 
@@ -170,72 +129,6 @@ def decp_processing(enable_cache_removal: bool = False):
         remove_unused_cache()
 
     print("☑️  Fin du flow principal decp_processing.")
-
-
-@task(log_prints=True)
-def enrich_from_sirene(df: pl.LazyFrame):
-    # Préprocessing des données SIRENE si :
-    # - le dossier n'existe pas encore (= les données n'ont pas déjà été preprocessed ce mois-ci)
-    # - on est au moins le 5 du mois (pour être sûr que les données SIRENE ont été mises à jour sur data.gouv.fr)
-    if not SIRENE_DATA_DIR.exists() and int(DATE_NOW[-2:]) >= 5:
-        sirene_preprocess()
-
-    # DONNÉES SIRENE ACHETEURS
-
-    print("Extraction des SIRET des acheteurs...")
-    df_sirets_acheteurs = extract_unique_acheteurs_siret(df.clone())
-
-    # print("Ajout des données établissements (acheteurs)...")
-    # df_sirets_acheteurs = add_etablissement_data(
-    #     df_sirets_acheteurs, ["enseigne1Etablissement"], "acheteur_id"
-    # )
-
-    print("Ajout des données unités légales (acheteurs)...")
-    df = add_unite_legale_data(
-        df, df_sirets_acheteurs, siret_column="acheteur_id", type_siret="acheteur"
-    )
-
-    # print("Construction du champ acheteur_nom à partir des données SIRENE...")
-    # df_sirets_acheteurs = make_acheteur_nom(df_sirets_acheteurs)
-
-    # print("Enregistrement des DECP aux formats CSV et Parquet...")
-    # save_to_files(df, f"{DIST_DIR}/decp")
-
-    # print("Suppression de colonnes et déduplication pour les DECP Sans Titulaires...")
-    # df_decp_sans_titulaires = make_decp_sans_titulaires(df)
-    # save_to_files(df_decp_sans_titulaires, f"{DIST_DIR}/decp-sans-titulaires")
-    # del df_decp_sans_titulaires
-
-    # DONNÉES SIRENE TITULAIRES
-
-    # Enrichissement des données pas prioritaire
-    # cf https://github.com/ColinMaudry/decp-processing/issues/17
-
-    print("Extraction des SIRET des titulaires...")
-    df_sirets_titulaires = extract_unique_titulaires_siret(df)
-
-    # print("Ajout des données établissements (titulaires)...")
-    # df_sirets_titulaires = add_etablissement_data_to_titulaires(df_sirets_titulaires)
-
-    print("Ajout des données unités légales (titulaires)...")
-    df = add_unite_legale_data(
-        df, df_sirets_titulaires, siret_column="titulaire_id", type_siret="titulaire"
-    )
-    # print("Amélioration des données unités légales des titulaires...")
-    # df_sirets_titulaires = improve_titulaire_unite_legale_data(df_sirets_titulaires)
-
-    # print("Renommage de certaines colonnes unités légales (titulaires)...")
-    # df_sirets_titulaires = rename_titulaire_sirene_columns(df_sirets_titulaires)
-
-    # print("Jointure pour créer les données DECP Titulaires...")
-    # df_decp_titulaires = merge_sirets_titulaires(df, df_sirets_titulaires)
-    # del df_sirets_titulaires
-
-    # print("Enregistrement des DECP Titulaires aux formats CSV et Parquet...")
-    # save_to_files(df_decp_titulaires, f"{DIST_DIR}/decp-titulaires")
-    # del df_decp_titulaires
-
-    return df
 
 
 @flow(log_prints=True)
@@ -247,7 +140,7 @@ def sirene_preprocess():
     print("🚀  Pré-traitement des données SIRENE")
     # Soit les tâches de ce flow vont au bout (success), soit le dossier SIRENE_DATA_DIR est supprimé (voir remove_sirene_data_dir())
     with transaction():
-        make_dirs_if_not_exist(SIRENE_DATA_DIR)
+        create_sirene_data_dir()
 
         # TODO préparer lest données établissements
